Trying again to include figures
I've had the devil of a time trying to get the figures to show up. I still don't know how to do it, but they are in the "figure" directory as instructed. If anybody can tell me why they're not showing up with the .md file or the .rmd file, please do so and thanks. I included a pdf file because that way you can see that I did create the figures.